refactor: removed old header
