+++ src/components/Header/styles.ts
@@ -1,45 +1,41 @@
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
 import theme from '../../styles/theme'
 
 export const Wrapper = styled.div`
-  position: relative;
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
 
   display: flex;
   justify-content: space-between;
   align-items: center;
 
   margin: 0 auto;
+  margin-top: 3.2rem;
   max-width: 114rem;
-  height: 11rem;
 
   z-index: ${theme.layers.menu};
-  @media (max-width: 1200px) {
-    padding: 0 3rem;
+  padding-inline: 2.4rem;
+
+  @media (max-width: 992px) {
+    margin-top: 2.4rem;
   }
 
-  @media (max-width: 540px) {
-    height: 8rem;
-    padding: 0 1.6rem;
+  @media (max-width: 576px) {
+    padding-inline: 1.6rem;
   }
 `
 
 export const LogoWrapper = styled.div`
   .logo-desktop {
-    @media (max-width: 960px) {
+    @media (max-width: 992px) {
       display: none;
     }
   }
 
   .logo-ipad {
-    > img {
-      width: 7rem;
-    }
-
-    @media (min-width: 961px) {
-      display: none;
-    }
-
-    @media (max-width: 539px) {
+    @media (min-width: 991px) {
       display: none;
     }
   }
@@ -49,155 +45,136 @@ export const LogoWrapper = styled.div`
 
 export const Menu = styled.nav`
   display: flex;
-  justify-content: space-between;
+  justify-content: flex-end;
   align-items: center;
+  gap: 3.2rem;
 
-  .logo-mobile {
-    width: 4.6rem;
-    margin-right: -0.8rem;
-
-    @media (min-width: 541px) {
-      display: none;
-    }
-    @media (max-width: 360px) {
-      width: 4.2rem;
-    }
-  }
-
-  @media (max-width: 960px) {
-    min-width: 58rem;
-  }
-
-  @media (max-width: 768px) {
-    min-width: 42rem;
-  }
-
-  @media (max-width: 540px) {
-    min-width: 100%;
+  @media (max-width: 992px) {
+    gap: 1.6rem;
   }
 `
+
+interface IMenuLinkProps {
+  active: boolean;
+}
 
 export const MenuLink = styled.a`
   position: relative;
 
-  margin-right: 2.4rem;
-  padding-top: 1.2rem;
-  padding-bottom: 1.3rem;
-
   color: ${theme.colors.snow};
   font-size: ${theme.font.sizes.font16};
-  font-weight: ${theme.font.weight.light};
   text-decoration: none;
   text-align: center;
 
   outline: none;
-  cursor: pointer;
 
-  &:hover,
-  &:focus-within {
-    &::after {
-      content: '';
-      position: absolute;
+  ${({ active }: IMenuLinkProps) => css`
+    font-weight: ${active
+      ? theme.font.weight.semibold
+      : theme.font.weight.light};
+  `}
 
-      display: block;
-      height: 0.3rem;
-      margin-top: 1.2rem;
-
-      background-color: ${theme.colors.cyan};
-      border-radius: 0.3rem;
-
-      animation: hoverAnimation 0.2s forwards;
-    }
-
-    @keyframes hoverAnimation {
-      from {
-        width: 0;
-        left: 50%;
-      }
-      to {
-        width: 100%;
-        left: 0;
-      }
-    }
-  }
-
-  @media (max-width: 540px) {
-    font-size: ${theme.font.sizes.font14};
-    margin-right: 0;
-  }
-`
-
-export const MenuLinkDisable = styled.a`
-  position: relative;
-
-  margin-right: 4.2rem;
-  padding-top: 1.2rem;
-  padding-bottom: 1.3rem;
-
-  color: ${theme.colors.grayDisabled};
-  font-size: ${theme.font.sizes.font16};
-  font-weight: ${theme.font.weight.light};
-  text-decoration: none;
-  text-align: center;
-
-  cursor: not-allowed;
-  &:hover {
-    &::after {
-      content: '';
-      position: absolute;
-
-      display: block;
-      height: 0.3rem;
-      margin-top: 1.2rem;
-
-      background-color: ${theme.colors.grayDisabled};
-      border-radius: 0.3rem;
-
-      animation: hoverAnimation 0.3s forwards;
-    }
-    @keyframes hoverAnimation {
-      from {
-        width: 0;
-        left: 50%;
-      }
-      to {
-        width: 100%;
-        left: 0;
-      }
-    }
-  }
-
-  @media (max-width: 768px) {
-    margin-right: 3.2rem;
-  }
-
-  @media (max-width: 541px) {
-    margin-right: 2rem;
-    font-size: ${theme.font.sizes.font14};
-  }
-
-  @media (max-width: 360px) {
-    margin-right: 1.4rem;
-    font-size: ${theme.font.sizes.font12};
-  }
-
-  img {
+  &::after {
+    content: '';
     position: absolute;
-    right: -2rem;
-    top: 1.8rem;
+    bottom: -1rem;
 
-    @media (max-width: 540px) {
-      right: -1.6rem;
-      top: 1.7rem;
-    }
-    @media (max-width: 360px) {
-      top: 1.6rem;
-    }
+    left: 50%;
+    width: 0;
+    height: 0.2rem;
+
+    background-color: ${theme.colors.cyan};
+    border-radius: 0.1rem;
+    box-shadow: 0 0 0.6rem ${theme.colors.cyan};
+
+    transition-duration: 300ms;
+    transition-timing-function: ease-in-out;
+    transition-property: width left;
+
+    ${({ active }: IMenuLinkProps) => css`
+      left: ${active ? '0' : '50%'};
+      width: ${active ? '100%' : '0'};
+    `}
+  }
+
+  &:hover::after {
+    left: 0%;
+    width: 100%;
+  }
+
+  @media (max-width: 576px) {
+    font-size: ${theme.font.sizes.font14};
   }
 `
+
+// export const MenuLinkDisable = styled.a`
+//   position: relative;
+
+//   margin-right: 4.2rem;
+//   padding-top: 1.2rem;
+//   padding-bottom: 1.3rem;
+
+//   color: ${theme.colors.grayDisabled};
+//   font-size: ${theme.font.sizes.font16};
+//   font-weight: ${theme.font.weight.light};
+//   text-decoration: none;
+//   text-align: center;
+
+//   cursor: not-allowed;
+
+//   &::after {
+//     content: '';
+//     position: absolute;
+//     bottom: 0;
+
+//     left: 50%;
+//     width: 0;
+//     height: 0.3rem;
+
+//     background-color: ${theme.colors.grayDisabled};
+//     border-radius: 0.15rem;
+
+//     transition-duration: 300ms;
+//     transition-timing-function: ease-in-out;
+//     transition-property: width left;
+//   }
+
+//   &:hover::after {
+//     left: 0%;
+//     width: 100%;
+//   }
+
+//   @media (max-width: 768px) {
+//     margin-right: 3.2rem;
+//   }
+
+//   @media (max-width: 541px) {
+//     margin-right: 2rem;
+//     font-size: ${theme.font.sizes.font14};
+//   }
+
+//   @media (max-width: 360px) {
+//     margin-right: 1.4rem;
+//     font-size: ${theme.font.sizes.font12};
+//   }
+
+//   img {
+//     position: absolute;
+//     right: -2rem;
+//     top: 1.8rem;
+
+//     @media (max-width: 540px) {
+//       right: -1.6rem;
+//       top: 1.7rem;
+//     }
+//     @media (max-width: 360px) {
+//       top: 1.6rem;
+//     }
+//   }
+// `
 
 export const MenuBottom = styled.div`
-  @media (max-width: 840px) {
+  @media (max-width: 576px) {
     position: fixed;
     right: 0;
     bottom: 0;
@@ -208,7 +185,6 @@ export const MenuBottom = styled.div`
     display: flex;
     justify-content: space-between;
     align-items: center;
-    // width: 100%;
     height: 6.8rem;
     padding: 1.6rem;
 
@@ -219,7 +195,11 @@ export const MenuBottom = styled.div`
     width: fit-content;
     padding: 1.2rem;
 
-    font-size: ${theme.font.sizes.font12};
+    font-size: ${theme.font.sizes.font14};
+
+    :hover {
+      font-weight: ${theme.font.weight.medium};
+    }
 
     img {
       width: 1.6rem;
@@ -229,16 +209,19 @@ export const MenuBottom = styled.div`
 
 export const ButtonsWrapper = styled.div`
   display: flex;
+  align-items: center;
   gap: 1rem;
-`
 
-export const OptionsContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  gap: 0.8rem;
+  a {
+    min-width: 12.5rem;
+    min-height: 3.8rem;
+  }
 
-  @media (min-width: 841px) {
-    display: none;
+  @media (max-width: 576px) {
+    flex-direction: row-reverse;
+    justify-content: space-between;
+
+    width: 100%;
   }
 `
 
@@ -247,12 +230,14 @@ export const ButtonOptions = styled.button`
   height: 3.2rem;
 
   background-color: rgba(255, 255, 255, 0.1);
-  border: none;
+  border: 0.1rem solid rgba(255, 255, 255, 0);
   border-radius: 50%;
-`
 
-export const MenuContainer = styled.div`
-  @media (max-width: 768px) {
-    display: none;
+  transition: border 300ms ease-in-out;
+
+  cursor: pointer;
+
+  &:hover {
+    border: 0.1rem solid rgba(255, 255, 255, 0.8);
   }
 `
